Use the ImageCacher to display exchange images

+++ mobile/BitcoinTicker/ImageCache.m
@@ -33,9 +33,9 @@
 
 + (NSString *)hash:(NSString *)string
 {
-    const char *cstr = [string UTF8String];
+    const char* cstr = [string UTF8String];
     unsigned char result[16];
-    CC_MD5(cstr, strlen(cstr), result);
+    CC_MD5(cstr, (int)strlen(cstr), result);
     return [NSString stringWithFormat:
             @"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
             result[0], result[1], result[2], result[3],
